requirements.txt
aiogram==3.13.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.1
redis==5.0.8
apscheduler==3.10.4
psycopg2-binary==2.9.9
